+++ lib/td_audit_web/views/notification_view.ex
@@ -1,0 +1,27 @@
+defmodule TdAuditWeb.NotificationView do
+  use TdAuditWeb, :view
+
+  alias TdAuditWeb.EventView
+
+  def render("index.json", %{notifications: notifications}) do
+    %{data: render_many(notifications, __MODULE__, "notifications.json")}
+  end
+
+  def render("notifications.json", %{notification: notification}) do
+    events =
+      notification.events
+      |> render_many(EventView, "event.json")
+      |> Enum.map(&with_name_and_path/1)
+
+    notification
+    |> Map.take([:id, :inserted_at])
+    |> Map.put(:events, events)
+  end
+
+  defp with_name_and_path(event) do
+    Map.merge(event, %{
+      name: EventView.resource_name(event),
+      path: EventView.path(event)
+    })
+  end
+end